♻️ refactor: Bookmark Test 분리 및 Place Test 수정
