+++ server/src/test/java/com/main21/bookmark/controller/BookmarkCreateTest.java
@@ -1,0 +1,52 @@
+package com.main21.bookmark.controller;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
+import org.springframework.http.MediaType;
+import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
+import org.springframework.test.web.servlet.ResultActions;
+
+import static com.main21.utils.AuthConstants.AUTHORIZATION;
+import static com.main21.utils.AuthConstants.REFRESH;
+import static org.mockito.BDDMockito.given;
+import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
+import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
+import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
+import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+public class BookmarkCreateTest extends BookmarkControllerTest {
+
+    @Test
+    @DisplayName("POST 북마크 생성 - 북마크가 존재하지 않으면 생성하기")
+    void createBookmark() throws Exception{
+
+        Long placeId = 1L;
+
+        given(redisUtils.getId(Mockito.anyString())).willReturn(1L);
+        given(bookmarkService.createBookmark(Mockito.anyLong(), Mockito.anyString())).willReturn(true);
+
+        ResultActions actions =
+                mockMvc.perform(
+                        RestDocumentationRequestBuilders.get("/bookmark/{place-id}", placeId)
+                                .header(AUTHORIZATION, "Bearer " + accessToken)
+                                .header(REFRESH, refreshToken)
+                                .accept(MediaType.APPLICATION_JSON)
+                                .contentType(MediaType.APPLICATION_JSON)
+                );
+
+        actions
+                .andExpect(status().isOk())
+                .andDo(document("북마크 추가",
+                        requestHeaders(
+                                headerWithName(AUTHORIZATION).description("엑세스 토큰"),
+                                headerWithName(REFRESH).description("리프레시 토큰")
+                        ),
+                        pathParameters(
+                                parameterWithName("place-id").description("장소 식별자")
+                        )
+                ));
+    }
+}
